Store Transactions: move payment creator functions to separate module
Let's not include the whole `lib/store-transactions` flow machinery into code
that just wants to set payment type.
