+++ client/my-sites/checkout/checkout/credit-card-selector.jsx
@@ -12,7 +12,7 @@ import { find, defer } from 'lodash';
 import analytics from 'lib/analytics';
 import CreditCard from 'components/credit-card';
 import NewCardForm from './new-card-form';
-import { newCardPayment, newStripeCardPayment, storedCardPayment } from 'lib/store-transactions';
+import { newCardPayment, newStripeCardPayment, storedCardPayment } from 'lib/transaction/payments';
 import { setPayment } from 'lib/transaction/actions';
 import { withStripeProps } from 'lib/stripe';
